feat: Component shared moved to shared path
