+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 import { RouterOutlet } from "@angular/router";
-import { NavBarComponent } from "./components/nav-bar/nav-bar.component";
-import { FooterComponent } from "./components/footer/footer.component";
+import { NavBarComponent } from "./components/shared/nav-bar/nav-bar.component";
+import { FooterComponent } from "./components/shared/footer/footer.component";
 
 @Component({
   selector: "app-root",
