+++ src/app/components/shared/footer/footer.component.less
