+++ src/app/components/shared/loading/loading.component.less
