+++ src/app/pages/home/home.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 import { AuthService } from "@auth0/auth0-angular";
-import { LoadingComponent } from "./../../components/loading/loading.component";
+import { LoadingComponent } from "../../components/shared/loading/loading.component";
 import { AsyncPipe, NgIf } from "@angular/common";
 import { ProductListComponent } from "../../components/product-list/product-list.component";
 import { MainBannerComponent } from "../../components/main-banner/main-banner.component";
